preps frame test
